+++ lib/main.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/widgets.dart';
 
 import 'src/app.dart';
+import 'src/deep_link_service.dart';
 import 'src/location_service.dart';
 import 'src/photo_service.dart';
 
@@ -10,6 +11,7 @@ Future<void> main() async {
     const GlassTrailBootstrapApp(
       photoService: FileSelectorPhotoService(),
       locationService: PlatformLocationService(),
+      deepLinkService: PlatformDeepLinkService(),
     ),
   );
 }